refactor(core): Split config.rs into domain modules
Refactors the 1779-line config.rs file into a modular directory
structure for better maintainability.

New structure:
- config/mod.rs - Main Config struct, re-exports (~690 lines)
- config/network.rs - NetworkConfig, BlobRegistryConfig, NatDialConfig
- config/observability.rs - ObservabilityConfig, TracingConfig, RateLimitingConfig
- config/supervisor.rs - SupervisorConfig, RestartPolicyConfig
- config/federation.rs - FederationConfig, FederationRetryConfig
- config/privacy.rs - PrivacyConfig
- config/steward.rs - StewardNodeConfig
- config/gateway.rs - GatewayConfig
- config/cooperative.rs - CooperativeConfig

Backwards compatible - all types re-exported from mod.rs so existing
imports continue to work unchanged.

Closes #566
