+++ icn/crates/icn-core/src/config/mod.rs
@@ -1,0 +1,783 @@
+//! Configuration management for ICNd
+//!
+//! This module provides configuration types for all ICN subsystems,
+//! organized by domain into sub-modules.
+
+mod cooperative;
+mod federation;
+mod gateway;
+mod network;
+mod observability;
+mod privacy;
+mod steward;
+mod supervisor;
+
+// Re-export all types for backwards compatibility
+pub use cooperative::*;
+pub use federation::*;
+pub use gateway::*;
+pub use network::*;
+pub use observability::*;
+pub use privacy::*;
+pub use steward::*;
+pub use supervisor::*;
+
+// Re-export topology types from icn-net to avoid circular dependencies
+pub use icn_net::{FanoutConfig, NeighborLimitsConfig, NodeRole, TopologyConfig};
+
+use serde::{Deserialize, Serialize};
+use std::path::PathBuf;
+
+/// ICNd configuration
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Config {
+    /// Data directory for ICN state
+    pub data_dir: PathBuf,
+
+    /// Network configuration
+    pub network: NetworkConfig,
+
+    /// Observability configuration
+    pub observability: ObservabilityConfig,
+
+    /// Rate limiting configuration
+    #[serde(default)]
+    pub rate_limiting: RateLimitingConfig,
+
+    /// Topology configuration for regional/cluster-based networking
+    #[serde(default)]
+    pub topology: TopologyConfig,
+
+    /// Gateway API configuration
+    #[serde(default)]
+    pub gateway: GatewayConfig,
+
+    /// Federation configuration for cross-network connectivity
+    #[serde(default)]
+    pub federation: FederationConfig,
+
+    /// Privacy configuration for metadata protection and onion routing
+    #[serde(default)]
+    pub privacy: PrivacyConfig,
+
+    /// Cooperative configuration for cooperative-specific settings
+    #[serde(default)]
+    pub cooperative: CooperativeConfig,
+
+    /// Steward configuration for SDIS steward network participation
+    #[serde(default)]
+    pub steward: StewardNodeConfig,
+
+    /// Supervisor configuration for background tasks and timeouts
+    #[serde(default)]
+    pub supervisor: SupervisorConfig,
+}
+
+impl Default for Config {
+    fn default() -> Self {
+        Config {
+            data_dir: dirs::data_dir()
+                .unwrap_or_else(|| PathBuf::from("."))
+                .join("icn"),
+            network: NetworkConfig {
+                mdns_enabled: true,
+                listen_addr: "0.0.0.0:7777".to_string(),
+                rpc_port: 5601,
+                bootstrap_peers: vec![],
+                stun_servers: default_stun_servers(),
+                min_trust_threshold: default_min_trust_threshold(),
+                turn_server: None,
+                turn_username: None,
+                turn_password: None,
+                e2e_encryption_enabled: true,
+                encryption_cleanup_circuit_breaker_threshold: 3,
+                nat_dial: NatDialConfig::default(),
+                blob_registry: BlobRegistryConfig::default(),
+            },
+            observability: ObservabilityConfig {
+                metrics_port: 9100,
+                health_port: 8080,
+                log_level: "info".to_string(),
+                tracing: TracingConfig::default(),
+            },
+            rate_limiting: RateLimitingConfig::default(),
+            topology: TopologyConfig::default(),
+            gateway: GatewayConfig::default(),
+            federation: FederationConfig::default(),
+            privacy: PrivacyConfig::default(),
+            cooperative: CooperativeConfig::default(),
+            steward: StewardNodeConfig::default(),
+            supervisor: SupervisorConfig::default(),
+        }
+    }
+}
+
+impl Config {
+    /// Load configuration from a TOML file
+    pub fn from_file(path: impl AsRef<std::path::Path>) -> anyhow::Result<Self> {
+        let contents = std::fs::read_to_string(path)?;
+        let config = toml::from_str(&contents)?;
+        Ok(config)
+    }
+
+    /// Save configuration to a TOML file
+    pub fn to_file(&self, path: impl AsRef<std::path::Path>) -> anyhow::Result<()> {
+        let contents = toml::to_string_pretty(self)?;
+        std::fs::write(path, contents)?;
+        Ok(())
+    }
+
+    /// Get the keystore path
+    pub fn keystore_path(&self) -> PathBuf {
+        self.data_dir.join("identity.age")
+    }
+
+    /// Get the store path
+    pub fn store_path(&self) -> PathBuf {
+        self.data_dir.join("store")
+    }
+
+    /// Validate configuration and return a list of warnings/errors
+    ///
+    /// Returns Ok(warnings) if config is valid (warnings are non-fatal),
+    /// Returns Err(errors) if config has fatal issues.
+    pub fn validate(&self) -> Result<Vec<String>, Vec<String>> {
+        let mut warnings = Vec::new();
+        let mut errors = Vec::new();
+
+        // Gateway validation
+        if self.gateway.enabled {
+            if self.gateway.jwt_secret.is_empty() {
+                errors.push(
+                    "Gateway is enabled but jwt_secret is empty. Set via:\n  \
+                     - Config: gateway.jwt_secret = \"your-secret\"\n  \
+                     - Environment: ICN_GATEWAY_JWT_SECRET\n  \
+                     - CLI: --gateway-jwt-secret"
+                        .to_string(),
+                );
+            } else if self.gateway.jwt_secret.len() < 32 {
+                warnings.push(format!(
+                    "Gateway jwt_secret is {} bytes, recommend at least 32 bytes for security",
+                    self.gateway.jwt_secret.len()
+                ));
+            }
+
+            if self.gateway.bind_addr.starts_with("0.0.0.0") {
+                warnings.push(
+                    "Gateway binding to 0.0.0.0 - ensure proper firewall/reverse proxy in production"
+                        .to_string(),
+                );
+            }
+        }
+
+        // Network validation
+        if self.network.listen_addr.is_empty() {
+            errors.push("network.listen_addr cannot be empty".to_string());
+        }
+
+        if self.network.min_trust_threshold < 0.0 || self.network.min_trust_threshold > 1.0 {
+            errors.push(format!(
+                "network.min_trust_threshold must be 0.0-1.0, got {}",
+                self.network.min_trust_threshold
+            ));
+        }
+
+        // Trust threshold warnings
+        if self.network.min_trust_threshold == 0.0 {
+            warnings.push(
+                "network.min_trust_threshold is 0.0 - trust-gated TLS is disabled, \
+                 all authenticated DIDs will be accepted"
+                    .to_string(),
+            );
+        }
+
+        // Federation validation
+        if self.federation.enabled {
+            if self.federation.coop_id.is_empty() {
+                warnings.push(
+                    "Federation enabled but coop_id is empty - will be derived from network_name"
+                        .to_string(),
+                );
+            }
+
+            if self.federation.bootstrap_peer_trust < 0.0
+                || self.federation.bootstrap_peer_trust > 1.0
+            {
+                errors.push(format!(
+                    "federation.bootstrap_peer_trust must be 0.0-1.0, got {}",
+                    self.federation.bootstrap_peer_trust
+                ));
+            }
+        }
+
+        // Privacy validation
+        if self.privacy.enabled && self.privacy.onion_routing_enabled {
+            if self.privacy.onion_hops < 2 {
+                warnings.push(format!(
+                    "privacy.onion_hops is {} - recommend at least 2 for meaningful anonymity",
+                    self.privacy.onion_hops
+                ));
+            }
+            if self.privacy.onion_hops > 5 {
+                warnings.push(format!(
+                    "privacy.onion_hops is {} - high latency expected, consider reducing",
+                    self.privacy.onion_hops
+                ));
+            }
+        }
+
+        // Steward validation
+        if self.steward.enabled {
+            if self.steward.vui_threshold > self.steward.vui_total_shares {
+                errors.push(format!(
+                    "steward.vui_threshold ({}) cannot exceed vui_total_shares ({})",
+                    self.steward.vui_threshold, self.steward.vui_total_shares
+                ));
+            }
+            if self.steward.vui_threshold < 2 {
+                warnings
+                    .push("steward.vui_threshold < 2 provides no threshold security".to_string());
+            }
+        }
+
+        // Rate limiting validation
+        if self.rate_limiting.enabled && self.rate_limiting.refill_interval_ms == 0 {
+            errors.push("rate_limiting.refill_interval_ms cannot be 0".to_string());
+        }
+
+        // Topology validation
+        if self.topology.fanout.local_cluster == 0 {
+            errors.push("topology.fanout.local_cluster cannot be 0".to_string());
+        }
+
+        if errors.is_empty() {
+            Ok(warnings)
+        } else {
+            Err(errors)
+        }
+    }
+
+    /// Print validation results to stderr with formatting
+    pub fn print_validation_results(&self) {
+        match self.validate() {
+            Ok(warnings) => {
+                for warning in warnings {
+                    eprintln!("\x1b[33mWarning:\x1b[0m {warning}");
+                }
+            }
+            Err(errors) => {
+                for error in &errors {
+                    eprintln!("\x1b[31mError:\x1b[0m {error}");
+                }
+                eprintln!(
+                    "\n\x1b[31mConfiguration has {} error(s). Fix them before starting.\x1b[0m",
+                    errors.len()
+                );
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_config_serialization() {
+        let config = Config::default();
+
+        // Serialize to TOML
+        let toml_str = toml::to_string_pretty(&config).unwrap();
+
+        // Should contain rate limiting section
+        assert!(toml_str.contains("[rate_limiting]"));
+        assert!(toml_str.contains("enabled"));
+        assert!(toml_str.contains("[rate_limiting.isolated]"));
+        assert!(toml_str.contains("[rate_limiting.known]"));
+        assert!(toml_str.contains("[rate_limiting.partner]"));
+        assert!(toml_str.contains("[rate_limiting.federated]"));
+        assert!(toml_str.contains("[rate_limiting.fallback]"));
+
+        // Deserialize back
+        let deserialized: Config = toml::from_str(&toml_str).unwrap();
+
+        // Verify rate limiting config
+        assert!(deserialized.rate_limiting.enabled);
+        assert_eq!(deserialized.rate_limiting.refill_interval_ms, 100);
+        assert_eq!(
+            deserialized.rate_limiting.isolated.max_messages_per_second,
+            10
+        );
+        assert_eq!(deserialized.rate_limiting.isolated.burst_capacity, 2);
+        assert_eq!(deserialized.rate_limiting.known.max_messages_per_second, 50);
+        assert_eq!(deserialized.rate_limiting.known.burst_capacity, 10);
+        assert_eq!(
+            deserialized.rate_limiting.partner.max_messages_per_second,
+            100
+        );
+        assert_eq!(deserialized.rate_limiting.partner.burst_capacity, 20);
+        assert_eq!(
+            deserialized.rate_limiting.federated.max_messages_per_second,
+            200
+        );
+        assert_eq!(deserialized.rate_limiting.federated.burst_capacity, 50);
+        assert_eq!(
+            deserialized.rate_limiting.fallback.max_messages_per_second,
+            100
+        );
+        assert_eq!(deserialized.rate_limiting.fallback.burst_capacity, 20);
+    }
+
+    #[test]
+    fn test_rate_limiting_config_conversion() {
+        let config = RateLimitingConfig::default();
+
+        // Convert to icn-net types with default trust threshold
+        let trust_gated = config.to_trust_gated_config(default_min_trust_threshold());
+        let fallback = config.to_fallback_config();
+
+        // Verify trust-gated config
+        assert_eq!(trust_gated.isolated.max_messages_per_second, 10);
+        assert_eq!(trust_gated.isolated.burst_capacity, 2);
+        assert_eq!(trust_gated.known.max_messages_per_second, 50);
+        assert_eq!(trust_gated.known.burst_capacity, 10);
+        assert_eq!(trust_gated.partner.max_messages_per_second, 100);
+        assert_eq!(trust_gated.partner.burst_capacity, 20);
+        assert_eq!(trust_gated.federated.max_messages_per_second, 200);
+        assert_eq!(trust_gated.federated.burst_capacity, 50);
+        assert_eq!(trust_gated.refill_interval.as_millis(), 100);
+        // Verify the passed trust threshold is used
+        assert!(
+            (trust_gated.min_trust_threshold - default_min_trust_threshold()).abs() < f64::EPSILON
+        );
+
+        // Verify fallback config
+        assert_eq!(fallback.max_messages_per_second, 100);
+        assert_eq!(fallback.burst_capacity, 20);
+        assert_eq!(fallback.refill_interval.as_millis(), 100);
+    }
+
+    #[test]
+    fn test_repository_config_files() {
+        // Test that the actual config files in the repository parse correctly
+        // CARGO_MANIFEST_DIR is always set during cargo test and points to the crate directory
+        let manifest_dir = std::env::var("CARGO_MANIFEST_DIR")
+            .expect("CARGO_MANIFEST_DIR should be set during cargo test");
+
+        // From icn-core crate: icn/crates/icn-core -> icn (workspace root) -> project root
+        let workspace_root = std::path::PathBuf::from(manifest_dir)
+            .parent() // -> crates/
+            .expect("manifest dir should have parent")
+            .parent() // -> icn/
+            .expect("crates dir should have parent")
+            .parent() // -> project root
+            .expect("workspace should have parent")
+            .to_path_buf();
+
+        let config_dir = workspace_root.join("config");
+
+        // Test icn-alpha.toml
+        let alpha_path = config_dir.join("icn-alpha.toml");
+        if alpha_path.exists() {
+            let alpha = Config::from_file(&alpha_path).unwrap();
+            assert_eq!(alpha.network.listen_addr, "0.0.0.0:7777");
+            assert_eq!(alpha.observability.metrics_port, 9100);
+            assert!(alpha.rate_limiting.enabled);
+            assert_eq!(alpha.rate_limiting.isolated.max_messages_per_second, 10);
+            assert_eq!(alpha.rate_limiting.federated.max_messages_per_second, 200);
+        }
+
+        // Test icn-beta.toml
+        let beta_path = config_dir.join("icn-beta.toml");
+        if beta_path.exists() {
+            let beta = Config::from_file(&beta_path).unwrap();
+            assert_eq!(beta.network.listen_addr, "0.0.0.0:7778");
+            assert_eq!(beta.observability.metrics_port, 9101);
+            assert!(beta.rate_limiting.enabled);
+        }
+
+        // Test icn.toml.example
+        let example_path = config_dir.join("icn.toml.example");
+        if example_path.exists() {
+            let example = Config::from_file(&example_path).unwrap();
+            assert!(example.rate_limiting.enabled);
+            assert_eq!(example.rate_limiting.partner.max_messages_per_second, 100);
+        }
+    }
+
+    #[test]
+    fn test_partial_rate_limiting_config() {
+        // Test that we can override individual rate limiting settings
+        let toml_str = r#"
+data_dir = "/tmp/icn"
+
+[network]
+mdns_enabled = true
+listen_addr = "0.0.0.0:7777"
+rpc_port = 5601
+bootstrap_peers = []
+
+[observability]
+metrics_port = 9100
+health_port = 8080
+log_level = "info"
+
+[rate_limiting]
+enabled = false
+refill_interval_ms = 200
+
+[rate_limiting.isolated]
+max_messages_per_second = 5
+burst_capacity = 3
+"#;
+
+        let config: Config = toml::from_str(toml_str).unwrap();
+
+        assert!(!config.rate_limiting.enabled);
+        assert_eq!(config.rate_limiting.refill_interval_ms, 200); // custom
+        assert_eq!(config.rate_limiting.isolated.max_messages_per_second, 5); // custom
+        assert_eq!(config.rate_limiting.isolated.burst_capacity, 3); // custom
+                                                                     // Other fields should use defaults
+        assert_eq!(config.rate_limiting.known.max_messages_per_second, 50);
+        assert_eq!(config.rate_limiting.known.burst_capacity, 10);
+    }
+
+    #[test]
+    fn test_gateway_config_defaults() {
+        let config = GatewayConfig::default();
+
+        assert!(!config.enabled); // Disabled by default
+        assert_eq!(config.bind_addr, "127.0.0.1:8080");
+        assert_eq!(config.token_expiry_hours, 24);
+        assert_eq!(config.challenge_ttl_minutes, 5);
+        assert_eq!(config.jwt_secret, "");
+    }
+
+    #[test]
+    fn test_gateway_config_serialization() {
+        let toml_str = r#"
+enabled = true
+bind_addr = "0.0.0.0:8080"
+token_expiry_hours = 48
+challenge_ttl_minutes = 10
+jwt_secret = "test-secret"
+"#;
+
+        let config: GatewayConfig = toml::from_str(toml_str).unwrap();
+
+        assert!(config.enabled);
+        assert_eq!(config.bind_addr, "0.0.0.0:8080");
+        assert_eq!(config.token_expiry_hours, 48);
+        assert_eq!(config.challenge_ttl_minutes, 10);
+        assert_eq!(config.jwt_secret, "test-secret");
+    }
+
+    #[test]
+    fn test_config_with_gateway() {
+        let config = Config::default();
+
+        // Verify gateway config is present and has defaults
+        assert!(!config.gateway.enabled);
+        assert_eq!(config.gateway.bind_addr, "127.0.0.1:8080");
+
+        // Serialize to TOML
+        let toml_str = toml::to_string_pretty(&config).unwrap();
+        assert!(toml_str.contains("[gateway]"));
+
+        // Deserialize back
+        let deserialized: Config = toml::from_str(&toml_str).unwrap();
+        assert!(!deserialized.gateway.enabled);
+        assert_eq!(deserialized.gateway.bind_addr, "127.0.0.1:8080");
+    }
+
+    #[test]
+    fn test_federation_config_defaults() {
+        let config = FederationConfig::default();
+
+        assert!(!config.enabled); // Disabled by default
+        assert_eq!(config.network_name, "icn-mainnet");
+        assert!((config.bootstrap_peer_trust - 0.3).abs() < f64::EPSILON);
+        assert!(!config.auto_accept_invites);
+        assert!((config.min_invite_trust - 0.5).abs() < f64::EPSILON);
+        assert_eq!(config.max_federations, 10);
+        assert!(config.announce_public_addr);
+
+        // Retry defaults
+        assert_eq!(config.retry.max_retries, 5);
+        assert_eq!(config.retry.initial_delay_secs, 1);
+        assert_eq!(config.retry.max_delay_secs, 60);
+        assert_eq!(config.retry.reconnect_interval_secs, 30);
+    }
+
+    #[test]
+    fn test_federation_config_serialization() {
+        let toml_str = r#"
+enabled = true
+network_name = "my-coop-network"
+bootstrap_peer_trust = 0.5
+auto_accept_invites = true
+min_invite_trust = 0.7
+max_federations = 5
+announce_public_addr = false
+
+[retry]
+max_retries = 10
+initial_delay_secs = 2
+max_delay_secs = 120
+reconnect_interval_secs = 60
+"#;
+
+        let config: FederationConfig = toml::from_str(toml_str).unwrap();
+
+        assert!(config.enabled);
+        assert_eq!(config.network_name, "my-coop-network");
+        assert!((config.bootstrap_peer_trust - 0.5).abs() < f64::EPSILON);
+        assert!(config.auto_accept_invites);
+        assert!((config.min_invite_trust - 0.7).abs() < f64::EPSILON);
+        assert_eq!(config.max_federations, 5);
+        assert!(!config.announce_public_addr);
+
+        // Retry settings
+        assert_eq!(config.retry.max_retries, 10);
+        assert_eq!(config.retry.initial_delay_secs, 2);
+        assert_eq!(config.retry.max_delay_secs, 120);
+        assert_eq!(config.retry.reconnect_interval_secs, 60);
+    }
+
+    #[test]
+    fn test_config_with_federation() {
+        let config = Config::default();
+
+        // Verify federation config is present and has defaults
+        assert!(!config.federation.enabled);
+        assert_eq!(config.federation.network_name, "icn-mainnet");
+
+        // Serialize to TOML
+        let toml_str = toml::to_string_pretty(&config).unwrap();
+        assert!(toml_str.contains("[federation]"));
+        assert!(toml_str.contains("[federation.retry]"));
+
+        // Deserialize back
+        let deserialized: Config = toml::from_str(&toml_str).unwrap();
+        assert!(!deserialized.federation.enabled);
+        assert_eq!(deserialized.federation.network_name, "icn-mainnet");
+    }
+
+    #[test]
+    fn test_partial_federation_config() {
+        // Test that we can override individual federation settings
+        let toml_str = r#"
+data_dir = "/tmp/icn"
+
+[network]
+mdns_enabled = true
+listen_addr = "0.0.0.0:7777"
+rpc_port = 5601
+bootstrap_peers = ["icn://did:icn:test@192.168.1.100:7777"]
+
+[observability]
+metrics_port = 9100
+health_port = 8080
+log_level = "info"
+
+[federation]
+enabled = true
+network_name = "test-network"
+bootstrap_peer_trust = 0.4
+"#;
+
+        let config: Config = toml::from_str(toml_str).unwrap();
+
+        // Custom federation settings
+        assert!(config.federation.enabled);
+        assert_eq!(config.federation.network_name, "test-network");
+        assert!((config.federation.bootstrap_peer_trust - 0.4).abs() < f64::EPSILON);
+
+        // Default federation settings
+        assert!(!config.federation.auto_accept_invites);
+        assert_eq!(config.federation.max_federations, 10);
+        assert_eq!(config.federation.retry.max_retries, 5);
+
+        // Verify bootstrap peers in network config
+        assert_eq!(config.network.bootstrap_peers.len(), 1);
+        assert!(config.network.bootstrap_peers[0].contains("did:icn:test"));
+    }
+
+    #[test]
+    fn test_steward_config_defaults() {
+        let config = StewardNodeConfig::default();
+
+        assert!(!config.enabled);
+        assert_eq!(config.vui_threshold, 3);
+        assert_eq!(config.vui_total_shares, 5);
+        assert_eq!(config.max_concurrent_enrollments, 100);
+        assert_eq!(config.max_concurrent_recoveries, 50);
+        assert_eq!(config.token_validity_secs, 7 * 24 * 60 * 60);
+    }
+
+    #[test]
+    fn test_steward_config_serialization() {
+        let toml_str = r#"
+enabled = true
+vui_threshold = 5
+vui_total_shares = 9
+max_concurrent_enrollments = 200
+max_concurrent_recoveries = 100
+token_validity_secs = 86400
+"#;
+
+        let config: StewardNodeConfig = toml::from_str(toml_str).unwrap();
+
+        assert!(config.enabled);
+        assert_eq!(config.vui_threshold, 5);
+        assert_eq!(config.vui_total_shares, 9);
+        assert_eq!(config.max_concurrent_enrollments, 200);
+        assert_eq!(config.max_concurrent_recoveries, 100);
+        assert_eq!(config.token_validity_secs, 86400);
+    }
+
+    #[test]
+    fn test_steward_config_conversion() {
+        let node_config = StewardNodeConfig {
+            enabled: true,
+            vui_threshold: 5,
+            vui_total_shares: 9,
+            max_concurrent_enrollments: 200,
+            max_concurrent_recoveries: 100,
+            token_validity_secs: 86400,
+        };
+
+        let steward_config = node_config.to_steward_config();
+
+        assert_eq!(steward_config.vui_threshold, 5);
+        assert_eq!(steward_config.vui_total_shares, 9);
+        assert_eq!(steward_config.max_concurrent_enrollments, 200);
+        assert_eq!(steward_config.max_concurrent_recoveries, 100);
+        assert_eq!(steward_config.token_validity_secs, 86400);
+    }
+
+    #[test]
+    fn test_config_with_steward() {
+        let config = Config::default();
+
+        // Verify steward config is present and has defaults
+        assert!(!config.steward.enabled);
+        assert_eq!(config.steward.vui_threshold, 3);
+        assert_eq!(config.steward.vui_total_shares, 5);
+
+        // Serialize to TOML
+        let toml_str = toml::to_string_pretty(&config).unwrap();
+        assert!(toml_str.contains("[steward]"));
+
+        // Deserialize back
+        let deserialized: Config = toml::from_str(&toml_str).unwrap();
+        assert!(!deserialized.steward.enabled);
+        assert_eq!(deserialized.steward.vui_threshold, 3);
+    }
+
+    #[test]
+    fn test_min_trust_threshold_config() {
+        // Test default value
+        let config = Config::default();
+        assert!((config.network.min_trust_threshold - 0.1).abs() < f64::EPSILON);
+
+        // Test serialization with default
+        let toml_str = toml::to_string_pretty(&config).unwrap();
+        assert!(toml_str.contains("min_trust_threshold"));
+
+        // Test custom value parsing
+        let custom_toml = r#"
+data_dir = "/tmp/icn"
+
+[network]
+mdns_enabled = true
+listen_addr = "0.0.0.0:7777"
+rpc_port = 5601
+bootstrap_peers = []
+min_trust_threshold = 0.0
+
+[observability]
+metrics_port = 9100
+health_port = 8080
+log_level = "info"
+"#;
+
+        let config: Config = toml::from_str(custom_toml).unwrap();
+        assert!((config.network.min_trust_threshold - 0.0).abs() < f64::EPSILON);
+
+        // Test that trust threshold is passed through to rate limiting config
+        let trust_gated = config
+            .rate_limiting
+            .to_trust_gated_config(config.network.min_trust_threshold);
+        assert!((trust_gated.min_trust_threshold - 0.0).abs() < f64::EPSILON);
+    }
+
+    #[test]
+    fn test_config_validation_default_passes() {
+        let config = Config::default();
+        let result = config.validate();
+        // Default config should pass validation (may have warnings)
+        assert!(result.is_ok());
+    }
+
+    #[test]
+    fn test_config_validation_gateway_jwt_required() {
+        let mut config = Config::default();
+        config.gateway.enabled = true;
+        config.gateway.jwt_secret = String::new();
+
+        let result = config.validate();
+        assert!(result.is_err());
+        let errors = result.unwrap_err();
+        assert!(errors.iter().any(|e| e.contains("jwt_secret")));
+    }
+
+    #[test]
+    fn test_config_validation_gateway_jwt_short_warning() {
+        let mut config = Config::default();
+        config.gateway.enabled = true;
+        config.gateway.jwt_secret = "short".to_string();
+
+        let result = config.validate();
+        assert!(result.is_ok());
+        let warnings = result.unwrap();
+        assert!(warnings.iter().any(|w| w.contains("32 bytes")));
+    }
+
+    #[test]
+    fn test_config_validation_steward_threshold() {
+        let mut config = Config::default();
+        config.steward.enabled = true;
+        config.steward.vui_threshold = 10;
+        config.steward.vui_total_shares = 5;
+
+        let result = config.validate();
+        assert!(result.is_err());
+        let errors = result.unwrap_err();
+        assert!(errors.iter().any(|e| e.contains("vui_threshold")));
+    }
+
+    #[test]
+    fn test_config_validation_trust_threshold_range() {
+        let mut config = Config::default();
+        config.network.min_trust_threshold = 1.5;
+
+        let result = config.validate();
+        assert!(result.is_err());
+        let errors = result.unwrap_err();
+        assert!(errors.iter().any(|e| e.contains("min_trust_threshold")));
+    }
+
+    #[test]
+    fn test_config_validation_zero_trust_warning() {
+        let mut config = Config::default();
+        config.network.min_trust_threshold = 0.0;
+
+        let result = config.validate();
+        assert!(result.is_ok());
+        let warnings = result.unwrap();
+        assert!(warnings
+            .iter()
+            .any(|w| w.contains("trust-gated TLS is disabled")));
+    }
+}
